Change data/items.json to data/items.js
This project should stay simple, so I am not willing to start
a HTTP server to store the mock json database. Reading the file
directly from the client is proving to be too much of a headache
(can't use fetch API, can't use XML requests, no way to read the
filesystem directly).

Therefore, I decided to hardcode the JSON data in a constant, that
will be stored in a file under the data directory.