Create header
Created the header, defined some styles and the model for the submenus
on each item of the nav. The submenus will be hidden / shown using
the CSS :hover selector.

The navs have transitions to switch the background color and the font
size on hover.

The filter input is already there, but not fully styled.