Finish the filter styling
Added a clean field button, which stays hidden unless there is some
text in the input. This effect was reached using only CSS, with the
:not(:placeholder-show) selector.

Also removed the border and outline of the filter input. Originally, I
planned to remove only the automatic blue borders that appears when the
input is focused or hovered, but no borders at all made it more alike to
the original site.